+++ packages/apps/src/app/trade/chart/query.ts
@@ -11,6 +11,7 @@ function getPriceQuery(range: string) {
 function getRange(range: ChartRange) {
   switch (range) {
     case ChartRange['1d']:
+      return `10m`;
     case ChartRange['1w']:
       return '1h';
     case ChartRange['1m']:
@@ -35,8 +36,8 @@ export function buildPriceQuery(
         block.height AS block,
         event.args->>'who' AS who,
         event.name AS operation,
-        token_metadata_in.symbol AS asset_in,
-        token_metadata_out.symbol AS asset_out,
+        (event.args->>'assetIn')::integer AS asset_in,
+        (event.args->>'assetOut')::integer AS asset_out,
         (event.args->>'amountIn')::numeric / (10 ^ token_metadata_in.decimals) AS amount_in,
         (event.args->>'amountOut')::numeric / (10 ^ token_metadata_out.decimals) AS amount_out
       FROM event
@@ -44,16 +45,16 @@ export function buildPriceQuery(
       INNER JOIN call ON call_id = call.id
       INNER JOIN token_metadata AS token_metadata_in ON (event.args->>'assetIn')::integer = token_metadata_in.id
       INNER JOIN token_metadata AS token_metadata_out ON (event.args->>'assetOut')::integer = token_metadata_out.id
-      WHERE call.name != 'Router.buy'
-        AND event.name IN ('Omnipool.BuyExecuted', 'Omnipool.SellExecuted', 'Router.RouteExecuted')
+      WHERE ((call.name != 'Router.buy' AND event.name = 'Router.RouteExecuted')
+        OR event.name IN ('Omnipool.BuyExecuted', 'Omnipool.SellExecuted', 'Router.Executed'))
         AND timestamp BETWEEN '${from}' AND '${to}'
     ),
     pair_price AS (
       SELECT 
         timestamp,
         CASE 
-          WHEN asset_in = '${assetIn}' AND asset_out = '${assetOut}' AND amount_in != 0 AND amount_out != 0 THEN amount_in / amount_out
-          WHEN asset_in = '${assetOut}' AND asset_out = '${assetIn}' AND amount_in != 0 AND amount_out != 0 THEN amount_out / amount_in
+          WHEN asset_in = ${assetIn} AND asset_out = ${assetOut} AND amount_in != 0 AND amount_out != 0 THEN amount_in / amount_out
+          WHEN asset_in = ${assetOut} AND asset_out = ${assetIn} AND amount_in != 0 AND amount_out != 0 THEN amount_out / amount_in
         END AS price
       FROM nor_trades
     )